Stress test the fast clear optimisation with multisample buffers
On Gen7+ in the i965 driver of Mesa there is an optimisation so that
when the color buffer is cleared with certain allowable colors a
separate smaller buffer is used to keep track of which pixels are
using this clear color and the clear color is stored outside the
buffer in the surface state. That way it doesn't have to actually
write the color to the color buffer. This adds a test which clears a
bunch of formats using fbo-formats.h and a few different clear colors.

In particular this demonstrates a bug currently in the i965 driver
where it wouldn't take into account missing components in the surface
so that clearing a GL_RED texture to white would return white rather
than red. It also demonstrates numerous problems with the
work-in-progress SKL implementation.

v2: Test all formats in fbo-formats.h as well as some different colors
    instead of just RGB and white.
Acked-by: Ben Widawsky <ben@bwidawsk.net>
Tested-by: Ben Widawsky <ben@bwidawsk.net>