ext_texture_array: Allow compressed test to run with GL_EXT_texture_compression_dxt1
All of the data is pre-compressed DXT1 textures, so there is no need for
any of the more functional S3TC related extensions.

Also... GL_ARB_texture_compress is a required part of OpenGL 1.3.  Since
the test already requires OpenGL 2.1, there is no reason to also check
for the extension.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Matt Turner <mattst88@gmail.com>