arb_get_texture_sub_image: Allow compressed test to run with GL_EXT_texture_compression_dxt1 and friends
All of the data is pre-compressed DXT[134] textures, so there is no need
for any of the more functional S3TC related extensions.

Also... GL_ARB_texture_non_power_of_two is a required part of OpenGL
2.0.  Since the test already requires OpenGL 2.0, there is no reason to
also check for the extension.

At some point someone should modify this test to exercise the
extension with the full set of compressed texture formats supported by
the implementation...

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Matt Turner <mattst88@gmail.com>