framework/test/glsl_parser_test.py: Handle gl versions correctly
This patch fixes the behavior of glsl_parser_test in cases other that
OpenGL and OpenGL ES are available. This means that if OpenGL ES isn't
available then OpenGL ES shaders will be passed to the regular version
of glslparsertest, which can run them with an ARB_ES<ver>_compatibility
extension. When a test is for desktop OpenGL, but only OpenGL ES is
available, then the test will skip in the python layer.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>

v3: - Rebase on top of now upstream fast-skipping patches
    - When using glslparsertest to run GLES tests add
      ARB_ES<ver>_compatibility to gl_required, to integrate with fast
      skipping.
    - Fix some tests. These tests are pretty fragile and probably just
      need to be reworked.
    - small refactors, make use of the _is_gles_version helper
    - Split skip conditions into a private method
v4: - Fix bug where ARB_ES<ver>_compatibility skip requirement would be
      overwritten if the test had other extension requirements (and
      update tests to catch this bug)
    - Rename some constants to be more clear
    - Remove use of flag for skipping.