framework: handle crash codes like piglit native tests.
This changes the behavior of the dEQP integration such that a status
that is < 0 on unix or (< 0 || == 3) on windows will be a crash rather
than a fail. A status > 0 (except 3 on windows) will still be marked
fail.

This makes use of the helper function from framework/test/base.py
rather than calling super() (which would still call the helper) because
we don't want to get the warn status that we would also inherit.

v2: - Update tests to use actual deqp ouput.
    - Fix status loop, v1 had a bug that would cause the loop to not
      exit when it needed to, but it would pass the simplified tests.
v3: - rename test helper function per Jason.
    - Simplify a massive if/elif statement in said test helper function

Tested-by: Jason Ekstrand <jason.ekstrand@intel.com> (v2)
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>