arb_internalformat_query2: add a generic check for TEXTURE_IMAGE_FORMAT
From spec:

"Possible values include any value that is legal to pass for the
 <format> parameter to the Tex*Image*D commands, or NONE if the
 resource is not supported for this operation."

Fails on NVIDIA proprietary drivers. In some cases is returns
GL_R11F_G11F_B10F or GL_RGB9_ES5, that are internalformat, not valid
formats.

Fails on ATI proprietary drivers, as it returns values different
to NONE on cases where the target/internalformat are not supported
(via INTERNALFORMAT_SUPPORTED).