arb_internalformat_query2: add generic pname check for GET_TEXTURE_IMAGE_FORMAT
From query2 spec:
"GET_TEXTURE_IMAGE_FORMAT:
 <skip>
 Possible values include any value that is legal to pass for the
 <format> parameter to GetTexImage, or NONE if the resource does not
 support this operation, or if GetTexImage is not supported."

The possible list of values is the same that testing
TEXTURE_IMAGE_FORMAT, but depending on the gl version,
GL_STENCIL_INDEX is allowed too. This commits adds another list of
possible values. We could have added a clone method that added extra
values, but for now that would be an overkill. We could add more if we
find more similar cases.

As with TEXTURE_IMAGE_FORMAT, this test fails on NVIDIA proprietary
drivers, as in some cases returns GL_R11F_G11F_B10F or GL_RGB9_E5
(that are internalformats, not formats)