arb_internalformat_query2: test for TEXTURE_COMPRESSED_BLOCK_<X>
This test add a check for the following pnames:
  * TEXTURE_COMPRESSED_BLOCK_WIDTH
  * TEXTURE_COMPRESSED_BLOCK_HEIGHT
  * TEXTURE_COMPRESSED_BLOCK_SIZE

On all those three, query2 spec says the following:
"If the internal format is not compressed, or the resource is not
 supported, 0 is returned."

We could have classified the existing internalformats on
compressed/non-compressed (similar to color-format/non-color-format
for COLOR_ENCONDING), but that seems pointless taking into account
that we already have TEXTURE_COMPRESSED to query if a internalformat
is compressed.

So this test queries TEXTURE_COMPRESSED and INTERNALFORMAT_SUPPORTED,
and if any of them is false, checks that the returned value for any of
those three pnames is zero.

Note that this make this test really similar to a generic-pname-check
for those three, but adding on top a TEXTURE_COMPRESSED check. This
make this test a candidate to be refactored. Not done here becuase it
will$

This test passes with NVIDIA proprietary drivers.