winsys-framework: Default to showing window
winsys-framework incorrectly assumed that it will always have ownership of the
pixels in its own buffers. If using the default framebuffer (i.e., when Piglit
is not running in FBO mode,) ownership will not be acquired unless the window is
mapped.

While this is not typically a problem because buffers are separate, NVIDIA's
Unified Back Buffer feature makes all windows share a back buffer while still
conforming to the OpenGL spec. If Piglit does not establish pixel ownership, its
output will be clobbered by other windows.

While this problem could be easily fixed by specifying PIGLIT_FORCE_WINDOW=1 or
-fbo, the current default takes advantage of undefined behavior in the OpenGL
spec. A better solution would be to replace PIGLIT_FORCE_WINDOW with
PIGLIT_NO_WINDOW, a flag that allows tests to be run without mapping a window or
using an FBO. The default behavior, the, would be to map a window. Really,
though, if users want to test with offscreen rendering, they should use FBOs
with flag -fbo.

More information on pixel ownership here:
https://www.opengl.org/documentation/specs/version1.1/glspec1.1/node94.html

Reviewed by: Arthur Huillet <ahuillet@nvidia.com>
Reviewed-by: Dave Airlie <airlied@redhat.com>