arb_internalformat_query2: Added tests for image textures
Checks that:
  * TEXEL_SIZE
  * IMAGE_COMPATIBILITY_CLASS
  * IMAGE_PIXEL_FORMAT
  * IMAGE_PIXEL_TYPE
return the values defined for them in Table 3.22 of the OpenGL 4.2
specification depending on the "Image Format" passed.

This test does not check unsupported or expected-failure cases, hence
it does not replace the "generic-pnames-check" test. We may want to
change that in the future to have everything in the same place.

This tests fails on NVIDIA proprietary drivers. In addition to the
errors already found on the generic-pnames test with this pnames, in
some cases NVIDIA implementation is returning values different to the
specified on Table 3.22.