fast-clear: Skip formats that are not color renderable
fbo-formats.h contains some formats that are not marked as color
renderable in the spec, eg, the non-sized formats such as RG_INTEGER.
The spec disallows these formats in glTexImage2DMultisample and it
generates an error. Previously this didn't matter because it would end
up making the FBO incomplete and the format would be skipped. However,
since 1f172a09ea2 there is an assert if it can't find any components
with bits in the texture and this gets hit for these invalid formats.
This patch makes it check for the GL_INVALID_ENUM error upfront after
calling glTexImage2DMultisample so that it skips the format earlier.

It could be nice to use GL_ARB_internalformat_query2 to check whether
the format is color renderable and avoid generating the error
entirely, but that extension isn't implemented yet on i965.

Cc: Vinson Lee <vlee@freedesktop.org>