Add a fast clear test for non-MSRT surfaces
ext_framebuffer_multisample-fast-clear can now take a parameter on the
command line to make it test a single-sample buffer instead. This is
worth testing at least on i965 because fast clears are handled
differently when multisampling is not used.

v2: Add comments for the sizes of the two textures (Topi Pohjolainen)
Reviewed-by: Topi Pohjolainen <topi.pohjolainen@intel.com>