tests/all.py: Make MSAA_SAMPLE_COUNTS contain str() instead of int()
In every case MSAA_SAMPLES_COUNTS are converted to str() anyway, (either
explicitly by calling str(), or implicitly by str.format(). This is
often done with a pattern like: (str(x) for x in MSAA_SAMPLE_COUNTS).

This is just a waste of time when the values could be strings from the
start.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>