framework/grouptools.py: Add a function for print formatting
This adds a function to grouptools that replaces grouptools.SEPARATOR
with '/'. This function is meant for use when printing to the console,
since most people find '/' easier to read than whatever is being used
internally to separate a group, and because consistence counts.

The implementation differs from that originally used in console_.py,
this implementation (tested on python 2.7.10) is roughly twice as fast.

v2: - fix typos

Reviewed-by Glenn Kennard <glenn.kennard@gmail.com>
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>