multisample-fast-clear: Test enabling GL_FRAMEBUFFER_SRGB
If ‘enable-fb-srgb’ is given on the command line to the fast clear
test it will now enable GL_FRAMEBUFFER_SRGB before clearing the
buffer. This should cause the clear color to be converted from SRGB to
linear before being written to the framebuffer. The expected color is
therefore converted to match. This exposes a bug on SKL in the i965
driver when clearing SRGB MSRTs.

Reviewed-by: Topi Pohjolainen <topi.pohjolainen@intel.com>