Test blending a fast clear color with GL_FRAMEBUFFER_SRGB enabled
On SKL in the i965 driver there is some special handling of the fast
clear optimisation when GL_FRAMEBUFFER_SRGB is enabled because the
hardware can't cope with the fast clear buffer in that case and the
color buffer needs to be resolved. This test just tries varies
different clear colors and enabling at GL_FRAMEBUFFER_SRGB at
different points to make sure this resolving works correctly.

v2: Add a comment about probing the left-hand side. Remove the outer
    loop iterating the two values of srgb_before_clear. (Topi
    Pohjolainen).
Reviewed-by: Topi Pohjolainen <topi.pohjolainen@intel.com>