framework/test/opengl.py: Fix opengl fast-skipping wflinfo errors
This extends the version handling methods of the WflInfo class to handle
two kinds of errors. First, it handles getting a patch version in OpenGL
and GLSL version strings. Second, it handles getting back
WFLINFO_GL_ERROR, such as is returned on the nvidia blob driver because
of the patch version.

v2: - update tests to cover more tests (Jose)
    - update gl_extensions to not return set(['WFLINFO_GL_ERROR']). This
      wouldn't break anything, but it wasn't strictly correct.

Reviewed-by: Jose Fonseca <jfonseca@vmware.com>
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>