Test calling glViewport while using EGL_KHR_surfaceless_context
The i965 driver currently has a bug where if no window system surface
is bound then it will crash when glViewport is called. This patch adds
a test case which just sets up a simple context and binds it with no
surface before calling glViewport.

As far as I can tell this is the first explicit test for surfaceless
contexts so it also adds the directory and the test group.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=93257
v2: Use piglit_require_egl_extension and eglGetDisplay with
    EGL_DEFAULT_DISPLAY. Minor style fixes. (Ian Romanick)