framework/test/base.py: Handle fail cases for tests.
I'm going to admit I'm a bit puzzled how this could have slipped through
without being caught (I'm guessing an unrelated change uncovered this).
But basically if a test doesn't raise an exception, and the returncode
is > 0, it should mark the test with a status of "fail", but it doesn't.
Instead the default status "notrun" is passed to the logger, which it
doesn't support, and an exception is raised. This results in the test
being reported as "incomplete", since the thread asserts before it
writes the new result.

This patch corrects that problem.

This fixes the following test on i965 from mesa master:
spec/arb_compute_shader/indirect-compute: incomplete -> fail

bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=93340
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Reviewed-by: Mark Janes <mark.a.janes@intel.com>