arb_internalformat_query2: query2 api-error check
Similar to the equivalent for arb_internalformat_query, but in
addition to test when a invalid pname/target combination should return
INVALID_ENUM, it also checks that a valid combination returns
NO_ERROR, and testing both GetInternalformativ and
GetInternalformati64v.

The rationale of this is that is really likely that the implementation
of arb_internal_format_query2 would reuse a lot of bits of
arb_internal_format_query, so we want to be sure that a combination
that was invalid with arb_internal_format_query is not considered
invalid by arb_internal_format_query2.

Tested on NVIDIA GeForce GTX 950 - NVIDIA 352.55: pass