arb_internalformat_query2: test for SAMPLES and NUM_SAMPLE_COUNTS pnames
For those pnames, it verify that the defined on the spec is fulfilled:
  NUM_SAMPLE_COUNTS
    <skip>
    If <internalformat> is not color-renderable, depth-renderable, or
    stencil-renderable (as defined in section 4.4.4), or if <target>
    does not support multiple samples (ie other than
    TEXTURE_2D_MULTISAMPLE, TEXTURE_2D_MULTISAMPLE_ARRAY, or
    RENDERBUFFER), 0 is returned.

  SAMPLES
    <skip>
    If <internalformat> is not color-renderable, depth-renderable, or
    stencil-renderable (as defined in section 4.4.4), or if <target>
    does not support multiple samples (ie other than
    TEXTURE_2D_MULTISAMPLE, TEXTURE_2D_MULTISAMPLE_ARRAY, or
    RENDERBUFFER), <params> is not modified.

This test is really similar to api-errors test at
arb_internalformat_query, with the difference that on that case an
error was defined, and here was "softed" to return some specific
values. It also covers query1 overrun test.

Those two tests fails on the following proprietary drivers:
  * 4.5.13399 on AMD Radeon (TM) R9 380 Series
  * 4.5.0 NVIDIA 352.55 on GeForce GTX 950/PCIe/SSE2

For NUM_SAMPLES_COUNTS, they report a value different of 0 (5 on the
NVIDIA, 4 on ATI) for target/internalformat that should return 0
(example: GL_TEXTURE_1D+GL_COMPRESSED_RGB).  It is worth to note that
for those cases, INTERNALFORMAT_SUPPORTED returns TRUE. It is not
clear if they provide multi-sample capabilities for those targets.

That seems to suggest that the paragraph was C&P from query1, and it
is being too restrictive, when the purpose of the query2 itself is "a)
provide a mechanism for implementations to declare support *above* the
minimum required by the specification".

For the SAMPLES case, it changes the first element of <params>,
for cases where it should remain unmodified.

Acked-by: Dave Airlie <airlied@redhat.com>