arb_internalformat_query2: check returned max-dimensions values against GetInteger
In addition to the previous checks, this checks that the returned
values when supported are the same that the ones you receive calling
GetIntegerv with equivalent pnames like GL_MAX_TEXTURE_SIZE,
GL_MAX_3D_TEXTURE_SIZE, etc.

All those are internal format-independent, meanwhile GetInternalformat
allows to specify the internal format. So in theory there is the
possibility of being different for some internal format. But in
practice, this is not happening on any driver at this moment. Query2
spec mentions this case:

    "7) There some <pnames> which it makes no sense to be qualified by
     a per-format/target scope, how should we handle them?
     e.g. MAX_WIDTH and MAX_HEIGHT might be the same for all formats.
     e.g. properties like AUTO_GENERATE_MIPMAP and
     MANUAL_GENERATE_MIPMAP might depend only on the GL version.

     <skip>

     A) Just use this entry point as is, if there are no per-format or
     target differences, it is perfectly acceptable to have the
     implementation return the same information for all valid
     parameters. This does allow implementations to report caveats
     that may exist for some formats but not others, even though all
     formats/targets may be supported."

So at this point, taking into account the current implementation, it
makes sense to check against those values. Probably in the future this
check needs to be removed.

Tested on NVIDIA GeForce GTX 950 - NVIDIA 352.55: pass

Tested on 4.5.13399 on AMD Radeon (TM) R9 380 Series: fails:
  * For MAX_WIDTH:
    * For 3d textures is returning MAX_TEXTURE_SIZE
    * For texture buffers is returning crap. Different values for
      each internalformat, including negative values.
  * For MAX_HEIGHT:
    * Ditto for 3d textures.
    * For RENDERBUFFER and TEXTURE_2D_MULTISAMPLE_ARRAY is returning
      0 even for supported values.

Anyway, take into account that even the basic dimension test (0 for
unsupported) was failing on ATI proprietary drivers.

Acked-by: Dave Airlie <airlied@redhat.com>