arb_internalformat_query2: add MAX_COMBINED_DIMENSIONS subtest to max-dimensions
From spec:

"MAX_COMBINED_DIMENSIONS: The maximum combined dimensions for the
  resource is returned in <params>. The combined dimensions is the
  product of the individual dimensions of the resource. For
  multisampled surfaces the number of samples is considered an
  additional dimension. Note that the value returned can be >= 2^32
  and should be queried with the 64-bit query.
  <skip>
  If the resource is unsupported, zero is returned."

In this case the returning value is a combination of the values of
MAX_WIDTH, MAX_HEIGHT, MAX_DEPTH, SAMPLES and the number of faces, so
the test can check that the returned value is correct in any case.

Note that there are cases that the combined value is greater that
2^32, but the test is also testing the 32-bit query. The spec doesn't
specify what the query should return in that case, so it is assumed
that any value would be correct.

Tested on NVIDIA GeForce GTX 950 - NVIDIA 352.55: fails for cube
map related targets. From spec:

"For cube map targets this is the maximum combined width, height and
 faces"

But NVIDIA propietary drivers is just combining width and height.