texwrap: fix stencil texturing a bit
Previously, the test was actually using all-zeros in the end for the image,
so unsurprisingly it passed, except when border color was used.
Add the special stencil case handling in a couple more cases.
This will still not test depth/stencil textures in stencil texturing mode, and
furthermore because the spec doesn't seem to say what should happen with all
the channels make it ignore all errors on non-first channel.

Reviewed-by: Brian Paul <brianp@vmware.com>