Replace vs-tes-tessinner-tessouter-inputs with two new tests.
(This is commit 42587e6d4bc8dd79be02b for the !TCS case.)

vs-tes-tessinner-tessouter-inputs tried to verify that all vector
components written by the TCS were present in the TES.  However, this
is not guaranteed.  According to the ARB_tessellation_shader spec:

"Tessellation Evaluation Shader Inputs
 [...]
 For triangular tessellation, gl_TessLevelOuter[3] and gl_TessLevelInner[1]
 will be undefined.  For isoline tessellation, gl_TessLevelOuter[2],
 gl_TessLevelOuter[3], and both values in gl_TessLevelInner[] are undefined."

This patch removes the broken test, and replaces it with two new ones.
One uses quads, and probes all 4/2 components.  The other uses
triangles, but only probes the 3/1 components that are actually defined.

Testing both types of domains is especially useful for i965, which
stores the data in different layouts depending on the domain.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>
Reviewed-by: Marek Olšák <marek.olsak@amd.com>