unittests: move framework/tests to unittests
Currently the framework is pretty confusing. There's 'test' which is a
package containing test classes, then there's 'tests' which contains
unit tests for the framework. This is obviously not optimal. Beyond that
the unittests get installed, and that isn't really necessary.

As a bonus this removes the unit tests from coverage reports, which is a
nice plus.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>