framework/exceptions.py: Drop PIGLIT_DEBUG
Previously a stack trace would be suppressed in all instances unless
the PIGLIT_DEBUG environment variable was truthy. This was suboptimal
for a number of reasons.

This patch removes that functionality, now only a subset of piglit
specific exceptions that are meant to be suppressed are. All other
exceptions are directly raised, stopping the runner immediately. Other
piglit specific exceptions have been extended to have custom error
messages rather than relying on the handler to add them.

v2: - remove unit tests that no longer apply

Reviewed-by: Jose Fonseca <jfonseca@vmware.com> (v1)
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>