Add an SSO test which mix-and-matches TCS/TES with different interfaces.
On i965, the TCS writes a memory location which the TES reads.  Both
shader stages need to agree upon the layout.  With SSO, this is tricky,
as OpenGL allows developers to freely mix-and-match separate TCS and TES
programs that have never been linked together.

To test this, we compile two TCS programs with different outputs, and
create SSO pipelines that use them with the same TES.  A layout
determined by the TES alone would not suffice, so this ensures that
the layout is updated when the TCS changes.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>
Reviewed-by: Timothy Arceri <timothy.arceri@collabora.com>