framework: record pid of test process in results
This information can be used on linux in conjunction with scripts to
associate information in dmesg with information in piglit results.

Adds the pid information to both the json and junit backends, and adds
tests for both.

This information is not published into any of the summary modes on
purpose. It's usefulness is limited to the system on which the test was
ran, and only while a log of dmesg exists.

v2: - remove extra newline in junit stderr (Mark)

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Reviewed-by: Mark Janes <mark.a.janes@intel.com>