glsl-1.50-geometry-end-primitive: ensure position data is reasonably aligned
The rect halves comparison is in fact not valid in general. The reason is that
the same geometry does not have the same precision even if it it just shifted
by some fixed amount in one direction.
As an example, some calculated x value near 7 will be near 263 if drawn with
a viewport x value of 256. The value near 7 has 5 bits more precision -
when calculating the fixed-point values for rasterization (with subpixel
accuracy), the lower value thus may be rounded in a different direction
than the higher one, even with correct rounding (not even entirely sure what
"correct" rounding here means, nearest-floor or nearest-even or whatever, the
problem is independent from that). This can then cause different pixels to be
covered by the primitive.
This causes a failure with this test in llvmpipe when the rounding mode is
changed slightly (from "mostly" nearest-away-from-zero to nearest-even). I was
not able to reproduce this with this test on nvidia or amd hardware, but they
are definitely affected in theory by the same issue (proven by some quick and
dirty test).
So, do floor(tmp*2048) / 2048 to ensure everything is reasonably aligned. This
still retains some subpixel bits (but only very few).

This may be a problem with more tests, albeit most tend to use vertex data
which is already sufficiently aligned (e.g. drawing simple aligned rects).

v2: use mul/floor/div sequence suggested by Ian Romanick instead of add/sub -
trying to beat the compiler may not be reliable (as it could reorder
operations).

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Jose Fonseca <jfonseca@vmware.com>