arb_internalformat_query2: Check if the resource is supported in max-dimensions
The ARB_internalformat_query2 specification says about
MAX_WIDTH, MAX_HEIGHT, MAX_DEPTH and MAX_COMBINED_DIMENSIONS <pnames>:

   "If the resource is unsupported, zero is returned."

From the same specification:
   "In the following descriptions, the term /resource/ is used
    to generically refer to an object of the appropriate type that has
    been created with <internalformat> and <target>."

We check if the /resource/ is supported by trying to create the object,
if an error is raised, we consider the /resource/ as "unsupported".
Before, we were only checking if the <internalformat> was supported
independently of the <target>.

Tested on NVIDIA GeForce GTX 950 - NVIDIA 352.55: after this commit,
all the subtests on this test fails.

Now that a stricter check for unsupported resources are in place, the
test fails on NVIDIA because for some unsupported resources it returns
a value different to zero.