arb_enhanced_layouts: UBO explicit offset test
This is just a lazy copy of the arb_uniform_buffer_object bufferstorage
test updated to make use of explicit offsets.

Test results:
Nvidia GeForce 840M - NVIDIA 352.41: fail

This fails on the Nvidia driver due to a bug in handling the named
uniform block in the fragment shader. If the name is removed the
test works as expected.