log_tests.py: fix incompatible seek for python 3
In python2 one can seek to negatives (-1 is the end of the file, for
example). In python 2.6+ (and exclusively in python 3) one instead uses
the second argument of seek to say whether to start at the start or end
of the file, and then an exact point. So file.seek(-1) -> file.seek(0,
2)

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Acked-by: Jose Fonseca <jfonseca@vmware.com>