arb_shading_language_420pack: test multiple layout qualifiers on the same line
Test results:
Nvidia GeForce 840M - NVIDIA 352.41:

The results from the Nvidia binary are a little all over the place with these
tests, most of the tests with arb_shading_language_420pack enabled fail but
if the GLSL version is changed to 4.20 they pass.
The only tests that pass with arb_shading_language_420pack enabled are the
global tests but that seems to be because they do not detect the duplicates
and fail the negative tests.

Mesa 11.2 (dev):
pass

V3: drop tests that should target ARB_enhanced_layouts
V2: fix version mismatch in location tests, and prefix variable with out
in global tests

Cc: Matt Turner <mattst88@gmail.com>