arb_shader_subroutine: Call a subroutine by the type name instead of by uniform name
This should be a compilation error because it shouldn't be able to find
a function prototype.  At least that's my understanding of the spec.

This was accidentally discovered while trying to reproduce bug #93722.

NOTE: This test fails on Mesa.

v2: Fix bad assignment to piglit_fragcolor.  Noticed by Ilia.  Also fix
the type of func_type to match the functions.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Ilia Mirkin <imirkin@alum.mit.edu> [v1]
Cc: Dave Airlie <airlied@redhat.com>