arb_enhanced_layouts: tests for duplicate layout identifiers
Test results:
Nvidia GeForce 840M - NVIDIA 352.41

The results from the Nvidia binary are a little all over the place with these
tests, most of the tests with arb_enhanced_layouts enabled fail but
if the GLSL version is changed to 4.20 they pass.
The only tests that pass with arb_enhanced_layouts enabled are the
global tests but that seems to be because they do not detect the duplicates
and fail the negative tests.

Reviewed-by: Matt Turner <mattst88@gmail.com>