piglit: change piglit command to python 3
This makes the piglit executable run with python 3 instead of python 2,
the legacy scripts (piglit-run, piglit-summary-html, etc) still run with
python 2.

The goal of this series has been to make python 2.7 a fallback rather
than a main option. If one wants to use python 2, it's a trivial change
to the python script to get that.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Acked-by: Jose Fonseca <jfonseca@vmware.com>