base_tests.py: make tests for timeout run on python 3.x
These tests are always valid on python 3, but only valid on python 2 if
subprocess32 is installed.

One unittest on python3 raises a warning. I've tracked it back to somewhere
in the python stdlib. I think that for some reason a file is being
unlinked twice, but I'm having trouble tracking it down, it works fine
under python 2.7, and I think it's safe to ignore.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Acked-by: Jose Fonseca <jfonseca@vmware.com>