arb_fb_no_attach: Test with an odd sample count
Adds a test that creates a framebuffer with no attachments and sets
the default sample count to 3. It then verifies that GL_SAMPLES for
the fbo is at least 3 and GL_SAMPLE_BUFFERS is 1. It also checks the
value of gl_NumSamples in the shader by writing to an SSBO with the
value to verify that it is the same as GL_SAMPLES.

Currently using an odd sample count with no attachments triggers an
assert in the i965 driver in Mesa.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=93957
Cc: Ilia Mirkin <imirkin@alum.mit.edu>
Reviewed-by: Ilia Mirkin <imirkin@alum.mit.edu>