completions/bash: Add a completion file for piglit
This file is a bash completion library for the 'piglit' command, it is
as complete and accurate as I can get it with my limited knowledge and
patience for bash.

It does require the _filedir function from bash_completions, which is
provided by most Linux distributions, the BSD's, and can be installed on
OSX via homebrew or macports. I made this decisions over using the
builtin features from bash because the builtins don't handle spaces in
file and folder names, and don't allow restrictions to be placed on the
type or extension like _filedir does.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>