Fix invalid glShaderSource call in pipeline_stats_comp.
The test was passing a pointer to an uninitialized integer value as the
glShaderSource length parameter.  This is meant to be an array
containing the length of each shader string.

Instead, just pass NULL, which makes the GL assume the strings are
null-terminated (which they are).

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=93335
Reviewed-by: Mark Janes <mark.a.janes@intel.com>