getteximage-formats: Add option to clear texture before rendering
Demonstrates a bug in Intel color resolve logic. If rendering to
texture is preceded by fast clear subsequent reads from the
texture require the underlying color buffer to be resolved. In
case of mipmap generation Intel driver needs to reallocate the
underlying buffer to fit the subsequent levels. However, copy of
the original level zero to newly allocated buffer is delayed until
the driver is about to generate level one using level zero as
source. This causes a color resolve of the original level zero.
Both the generation of level one and the resolve are meta
operations, and the latter will interfere with the sampler
settings of the former.

Signed-off-by: Topi Pohjolainen <topi.pohjolainen@intel.com>
Reviewed-by: Brian Paul <brianp@vmware.com>