tox.ini: Set the mako version to the minimum version in cmake
We actually define a minimum required mako version in cmake. Since tox
is supposed to be useful for testing, we should test that version.

Mako 1.0.2 contains bug fixes for python 3.5. For versions of python
earlier than 3.5 0.8.0, is still sufficient, but for 3.5 1.0.2 is needed.

CMake is updated to check this as well.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>