arb_enhanced_layouts: xfb_stride compiler test fixes
Make sure to call check_link when error can be a compile or link time
error. Also fix a couple of bugs 1) xfb_stride should be compared to
max interleaved components multiplied by four 2) in the type tests
xfb_stride must be the same for each buffer so add a different buffer
to each declaration.