transform_feedback: extend testing of generated names for api matching
This tests an interface block array which has and array as a member,
importantly the members array size is large than the blocks array size.

This exposes a bug in Mesa where we incorrectly use the blocks array
size as well as the members after lowering the named interface block.

So Mesa thinks a valid name is:

WithInstArray[0].g[0][0] when it should be WithInstArray[0].g[0]

Running this on the Nvidia blob produces an error message which indicates
they have a similar bug:

   "Failed to link: error: Transform feedback varying WithInstArray[0].g[2]
   has index 2, but the array size is 2."