gen_flat_interpolation_qualifier.py: Adds non-flat interpolation tests generator
Generator for error checking on "flat" keyword.

This generator adds, additionally, checks for variables inside structs
and interface blocks, which weren't explicitly mentioned in the GLSL
specs and, partially, in the GLSL ES specs.

For a discussion about this check
https://lists.freedesktop.org/archives/mesa-dev/2016-March/109117.html
and Khronos bug #15671.

Also, removed 8 redundant tests replaced by the generator.

Reviewed-by: Dylan Baker <baker.dylan.c@gmail.com>
Acked-by: Dave Airlie <airlied@redhat.com>
Signed-off-by: Andres Gomez <agomez@igalia.com>