framework: Add a test runner for the Vulkan CTS
This is mostly a copy-and-paste of the original GLES3 dEQP runner with a
few added tidbits that makes running the Vulkan CTS easier.  The two
force-skip cases were very useful in the early days when lots of tests were
still in development and may no longer be needed.  However, I have a
feeling that we will yet want them.

v2 (Dylan):
 - Fix PEP8 issues
 - add and update docstrings
 - update extra_args for global deqp changes
 - update piglit.conf.example
 - remove unused functions
 - use __future__
 - use "bin" instead of "exe" (gles31 is weird that it uses "exe" not
   "bin", all other deqp modules use "bin")

Reviewed-by: Dylan Baker <dylanx.c.baker@intel.com>
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>