framework: Add handler for failure and error when expecting the other
Right now the JUnit backend knows what to do with a failure when it
expects one (or a crash), but not what to do when it expects a failure
and gets a crash (or vice versa).

This patch teaches it what to do in that case, mark the test as a
failure and give a message explaining why.

v2: - fix report for expect failure got crash (Mark)

signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Acked-by: Mark Janes <mark.a.janes@intel.com>